uml model classes completed at atribute level

+++ src/main/java/ar/edu/unq/misCursosUNQ/Repos/CourseRepo.java
@@ -6,7 +6,7 @@ import org.springframework.stereotype.Repository;
 import ar.edu.unq.misCursosUNQ.Course;
 
 @Repository
-public interface CourseRepo extends JpaRepository<Course, Long> {
+public interface CourseRepo extends JpaRepository<Course, Integer> {
 	/* By this simple extension, MateriaRepo inherits several methods 
 	 * for working with Materia persistence, including methods for saving, deleting, and finding Materia entities.
 	 * Along with default provided methods, we can add our own custom methods and queries to this interface. */
